refactor feature steps
