+++ features/steps/dashboard/dashboard_merge_requests.rb
@@ -1,14 +1,13 @@
 class DashboardMergeRequests < Spinach::FeatureSteps
+  include SharedAuthentication
+  include SharedPaths
+
   Then 'I should see my merge requests' do
     merge_requests = @user.merge_requests
     merge_requests.each do |mr|
       page.should have_content(mr.title[0..10])
       page.should have_content(mr.project.name)
     end
-  end
-
-  Given 'I sign in as a user' do
-    login_as :user
   end
 
   And 'I have authored merge requests' do
@@ -18,16 +17,7 @@ class DashboardMergeRequests < Spinach::FeatureSteps
     project1.add_access(@user, :read, :write)
     project2.add_access(@user, :read, :write)
 
-    merge_request1 = Factory :merge_request,
-      :author => @user,
-      :project => project1
-
-    merge_request2 = Factory :merge_request,
-      :author => @user,
-      :project => project2
-  end
-
-  And 'I visit dashboard merge requests page' do
-    visit dashboard_merge_requests_path
+    merge_request1 = Factory :merge_request, :author => @user, :project => project1
+    merge_request2 = Factory :merge_request, :author => @user, :project => project2
   end
 end
