+++ features/steps/project/project.rb
@@ -1,0 +1,5 @@
+class Projects < Spinach::FeatureSteps
+  include SharedAuthentication
+  include SharedProject
+  include SharedPaths
+end
