+++ features/steps/project/project_comment_commit.rb
@@ -1,0 +1,6 @@
+class ProjectCommentCommit < Spinach::FeatureSteps
+  include SharedAuthentication
+  include SharedProject
+  include SharedNote
+  include SharedPaths
+end
