+++ features/steps/project/project_issues.rb
@@ -1,4 +1,9 @@
 class ProjectIssues < Spinach::FeatureSteps
+  include SharedAuthentication
+  include SharedProject
+  include SharedNote
+  include SharedPaths
+
   Given 'I should see "Release 0.4" in issues' do
     page.should have_content "Release 0.4"
   end
@@ -51,20 +56,6 @@ class ProjectIssues < Spinach::FeatureSteps
     page.should have_content issue.project.name
   end
 
-  Given 'I visit issue page "Release 0.4"' do
-    issue = Issue.find_by_title("Release 0.4")
-    visit project_issue_path(issue.project, issue)
-  end
-
-  And 'I leave a comment like "XML attached"' do
-    fill_in "note_note", :with => "XML attached"
-    click_button "Add Comment"
-  end
-
-  Then 'I should see comment "XML attached"' do
-    page.should have_content "XML attached"
-  end
-
   Given 'I fill in issue search with "Release"' do
     fill_in 'issue_search', with: "Release"
   end
@@ -90,22 +81,14 @@ class ProjectIssues < Spinach::FeatureSteps
     project = Project.find_by_name("Shop")
     milestone = Factory :milestone, :title => "v2.2", :project => project
 
-    3.times do
-      issue = Factory :issue, :project => project, :milestone => milestone
-    end
+    3.times { Factory :issue, :project => project, :milestone => milestone }
   end
 
   And 'project "Shop" has milestone "v3.0"' do
     project = Project.find_by_name("Shop")
     milestone = Factory :milestone, :title => "v3.0", :project => project
 
-    3.times do
-      issue = Factory :issue, :project => project, :milestone => milestone
-    end
-  end
-
-  And 'I visit project "Shop" issues page' do
-    visit project_issues_path(Project.find_by_name("Shop"))
+    3.times { Factory :issue, :project => project, :milestone => milestone }
   end
 
   When 'I select milestone "v3.0"' do
@@ -130,15 +113,6 @@ class ProjectIssues < Spinach::FeatureSteps
     project = Project.find_by_name "Shop"
     assignee_name = project.users.first.name
     page.find(issues_assignee_selector).should have_content(assignee_name)
-  end
-
-  Given 'I sign in as a user' do
-    login_as :user
-  end
-
-  And 'I own project "Shop"' do
-    @project = Factory :project, :name => "Shop"
-    @project.add_access(@user, :admin)
   end
 
   And 'project "Shop" have "Release 0.4" open issue' do
