+++ features/steps/project/project_merge_requests.rb
@@ -1,4 +1,9 @@
 class ProjectMergeRequests < Spinach::FeatureSteps
+  include SharedAuthentication
+  include SharedProject
+  include SharedNote
+  include SharedPaths
+
   Then 'I should see "Bug NS-04" in merge requests' do
     page.should have_content "Bug NS-04"
   end
@@ -56,29 +61,6 @@ class ProjectMergeRequests < Spinach::FeatureSteps
     page.should have_content "Wiki Feature"
   end
 
-  Given 'I visit merge request page "Bug NS-04"' do
-    mr = MergeRequest.find_by_title("Bug NS-04")
-    visit project_merge_request_path(mr.project, mr)
-  end
-
-  And 'I leave a comment like "XML attached"' do
-    fill_in "note_note", :with => "XML attached"
-    click_button "Add Comment"
-  end
-
-  Then 'I should see comment "XML attached"' do
-    page.should have_content "XML attached"
-  end
-
-  Given 'I sign in as a user' do
-    login_as :user
-  end
-
-  And 'I own project "Shop"' do
-    @project = Factory :project, :name => "Shop"
-    @project.add_access(@user, :admin)
-  end
-
   And 'project "Shop" have "Bug NS-04" open merge request' do
     project = Project.find_by_name("Shop")
     Factory.create(:merge_request,
@@ -94,9 +76,5 @@ class ProjectMergeRequests < Spinach::FeatureSteps
       :project => project,
       :author => project.users.first,
       :closed => true)
-  end
-
-  And 'I visit project "Shop" merge requests page' do
-    visit project_merge_requests_path(Project.find_by_name("Shop"))
   end
 end
