+++ features/steps/project/project_milestones.rb
@@ -1,4 +1,8 @@
 class ProjectMilestones < Spinach::FeatureSteps
+  include SharedAuthentication
+  include SharedProject
+  include SharedPaths
+
   Then 'I should see milestone "v2.2"' do
     milestone = @project.milestones.find_by_title("v2.2")
     page.should have_content(milestone.title[0..10])
@@ -26,26 +30,10 @@ class ProjectMilestones < Spinach::FeatureSteps
     page.should have_content("Browse Issues")
   end
 
-  Given 'I sign in as a user' do
-    login_as :user
-  end
-
-  And 'I own project "Shop"' do
-    @project = Factory :project, :name => "Shop"
-    @project.add_access(@user, :admin)
-  end
-
   And 'project "Shop" has milestone "v2.2"' do
     project = Project.find_by_name("Shop")
     milestone = Factory :milestone, :title => "v2.2", :project => project
 
-    3.times do
-      issue = Factory :issue, :project => project, :milestone => milestone
-    end
-  end
-
-  Given 'I visit project "Shop" milestones page' do
-    @project = Project.find_by_name("Shop")
-    visit project_milestones_path(@project)
+    3.times { Factory :issue, :project => project, :milestone => milestone }
   end
 end
