+++ features/steps/project/project_wall.rb
@@ -1,0 +1,6 @@
+class ProjectWall < Spinach::FeatureSteps
+  include SharedAuthentication
+  include SharedProject
+  include SharedNote
+  include SharedPaths
+end
