+++ features/steps/shared/authentication.rb
@@ -1,0 +1,10 @@
+require Rails.root.join('spec', 'support', 'login_helpers')
+
+module SharedAuthentication
+  include Spinach::DSL
+  include LoginHelpers
+
+  Given 'I sign in as a user' do
+    login_as :user
+  end
+end
