+++ features/steps/shared/project.rb
@@ -1,0 +1,8 @@
+module SharedProject
+  include Spinach::DSL
+
+  And 'I own project "Shop"' do
+    @project = Factory :project, :name => "Shop"
+    @project.add_access(@user, :admin)
+  end
+end
